// src/utils/enums/envConfig.enum.ts
export enum EnvConfigEnum {
  APP_AWS_REGION = 'APP_AWS_REGION',
  APP_AWS_ACCESS_KEY_ID = 'APP_AWS_ACCESS_KEY_ID',
  APP_AWS_SECRET_ACCESS_KEY = 'APP_AWS_SECRET_ACCESS_KEY',
  APP_AWS_BUCKET = 'APP_AWS_BUCKET',
  PORT = 'PORT',
  TOKEN_SECRET = 'TOKEN_SECRET',
  REFRESH_TOKEN_SECRET = 'REFRESH_TOKEN_SECRET',
  CONNECTION_STRING = 'CONNECTION_STRING',
  TOKEN_EXPIRATION_TIME = 'TOKEN_EXPIRATION_TIME',
  REFRESH_TOKEN_EXPIRATION_TIME = 'REFRESH_TOKEN_EXPIRATION_TIME',
  RESET_PASSWORD_TOKEN_EXPIRATION_TIME = 'RESET_PASSWORD_TOKEN_EXPIRATION_TIME',
  GOOGLE_CLIENT_ID = 'GOOGLE_CLIENT_ID',
  GOOGLE_CLIENT_SECRET = 'GOOGLE_CLIENT_SECRET',
  GOOGLE_URL = 'GOOGLE_URL',
  GOOGLE_REDIRECT_URL = 'GOOGLE_REDIRECT_URL',
  REDIRECT_URI= 'REDIRECT_URI',
  FACEBOOK_CLIENT_ID='FACEBOOK_CLIENT_ID',
  FACEBOOK_CLIENT_SECRET = 'FACEBOOK_CLIENT_SECRET',
  FACEBOOK_SCOPE = 'FACEBOOK_SCOPE',
  FACEBOOK_REDIRECT_URI = 'FACEBOOK_REDIRECT_URI',
  SESSION_SECRET_KEY='SESSION_SECRET_KEY',
  GOOGLE_API_KEY= 'GOOGLE_API_KEY'
}